requirements.txt
gradio>=4.0.0
langchain>=0.1.0
langchain-community
langchain-chroma
langchain-mistralai>=0.1.0
chromadb>=0.4.0
mistralai>=0.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
# Orchestration
prefect>=2.14.0
prefect-shell>=0.2.0
# Monitoring et testing
psutil>=5.9.0
python-magic>=0.4.27
requests>=2.31.0
# Performance
aiofiles>=23.0.0